requirements.txt
Django==1.11.5
django-appconf==1.0.2
django-countries==4.6.2
django-filter==1.0.4
django-staticfiles==1.2.1
djangorestframework==3.6.4
Markdown==2.6.9
pytz==2017.2
